# project/api_documentation.yaml
openapi: 3.0.3
info:
  title: EASM - External Attack Surface Management API
  description: |
    Comprehensive API for external attack surface management including domain scanning,
    vulnerability assessment, and security monitoring capabilities.

    ## Features
    - Domain and subdomain enumeration
    - SSL/TLS certificate analysis  
    - DNS record analysis
    - Vulnerability scanning
    - Batch processing capabilities
    - Real-time monitoring and health checks
    - Caching and optimization

    ## Rate Limiting
    - Default: 100 requests per hour per IP
    - Authenticated users: 1000 requests per hour
    - Bulk operations: 10 requests per hour

    ## Authentication
    Currently supports API key authentication. JWT tokens will be supported in future versions.

  version: 2.0.0
  contact:
    name: EASM Development Team
    email: security@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.easm.example.com
    description: Production server
security:
  - ApiKeyAuth: []
paths:
...
  /scan:
    post:
      tags:
        - Scanning
      summary: Scan a single domain
      description: Performs comprehensive security scan on a single domain
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                domain:
                  type: string
                  pattern: '^[a-zA-Z0-9][a-zA-Z0-9\-\.]*[a-zA-Z0-9]$'
                  example: "example.com"
                  description: Domain name to scan
                scan_types:
                  type: array
                  items:
                    type: string
                    enum:
                      [dns, ssl, vulnerabilities, subdomains, related_domains]
                  example: ["dns", "ssl", "vulnerabilities"]
                  description: Types of scans to perform
              required:
                - domain
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "200":
          description: Scan completed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /history:
    get:
      tags:
        - History
      summary: Get scan history
      description: Returns paginated list of previous scans
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - name: domain
          in: query
          schema:
            type: string
          description: Filter by domain name
        - name: scan_type
          in: query
          schema:
            type: string
            enum: [dns, ssl, vulnerabilities, subdomains, related_domains]
          description: Filter by scan type
      responses:
        "200":
          description: Scan history retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanHistory"
...
  /api/tasks:
    get:
      tags:
        - Tasks
      summary: Get background tasks
      description: Returns list of background tasks with filtering and pagination
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
          description: Filter by task status
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Task list retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskList"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ScanRequest:
      type: object
      properties:
        domain:
          type: string
          pattern: '^[a-zA-Z0-9][a-zA-Z0-9\-\.]*[a-zA-Z0-9]$'
          example: "example.com"
        scan_types:
          type: array
          items:
            type: string
            enum: [dns, ssl, vulnerabilities, subdomains, related_domains]
          example: ["dns", "ssl", "vulnerabilities"]
      required:
        - domain
    ScanResult:
      type: object
      properties:
        scan_id:
          type: integer
        domain:
          type: string
        scan_date:
          type: string
          format: date-time
        scan_types:
          type: array
          items:
            type: string
        results:
          type: object
          properties:
            dns_records:
              type: object
            ssl_info:
              type: object
            vulnerabilities:
              type: array
              items:
                type: object
            subdomains:
              type: array
              items:
                type: string
            related_domains:
              type: array
              items:
                type: string
        metadata:
          type: object
          properties:
            scan_duration_seconds:
              type: number
            cached:
              type: boolean
            cache_hit_rate:
              type: number
...
    ScanHistory:
      type: object
      properties:
        total_count:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
        scans:
          type: array
          items:
            type: object
            properties:
              scan_id:
                type: integer
              domain:
                type: string
              scan_date:
                type: string
                format: date-time
              scan_types:
                type: array
                items:
                  type: string
              status:
                type: string
              duration_seconds:
                type: number
...
    TaskList:
      type: object
      properties:
        total_count:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskSummary"
    TaskSummary:
      type: object
      properties:
        task_id:
          type: string
        task_type:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        priority:
          type: integer
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        progress:
          type: number
          minimum: 0
          maximum: 100
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
...